.scripts/requirements.txt
python-frontmatter==1.0.0
pytz==2022.1
requests==2.27.1